+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### [Campaign]
 
 - Added `dimensions` of banner into required inputs during the creation of custom HTML banner. remp/remp#1165
+- Added operator `is not` for the pageview attributes of campaign settings. remp/remp#1177
 
 ### [Mailer]